Revert "* fixed: 在OSX/iOS环境下，内存地址转换成整形数据会出现编译错误（警告）。需要转换成专门的特殊整形数据类型."
This reverts commit dd7bcb38ee8183a2a6cb0a6605e5d1f898438c51.

+++ code/engine/gkCommon/IResource.h
@@ -125,7 +125,7 @@ public:
 #ifdef __x86_64__
         m_resLock = new gkScopedLock<gkMutexLock>( eLGID_Resource, (uint64)this );
 #else
-		m_resLock = new gkScopedLock<gkMutexLock>( eLGID_Resource, (uintptr_t)this );
+		m_resLock = new gkScopedLock<gkMutexLock>( eLGID_Resource, (uint32)this );
 #endif
